added a test which fails due to our fix from project and the correspoding fix
